Consolidate the naming of test classes
