+++ Tests/Testably.Abstractions.Testing.Tests/Helpers/EncryptionHelperTests.cs
@@ -1,6 +1,6 @@
 ﻿using Testably.Abstractions.Testing.Helpers;
 
-namespace Testably.Abstractions.Testing.Tests.Internal;
+namespace Testably.Abstractions.Testing.Tests.Helpers;
 
 public class EncryptionHelperTests
 {
