+++ Tests/Testably.Abstractions.Testing.Tests/Helpers/FilePlatformIndependenceExtensionsTests.cs
@@ -2,7 +2,7 @@ using System.IO;
 using Testably.Abstractions.Testing.Helpers;
 using Testably.Abstractions.Testing.Tests.TestHelpers;
 
-namespace Testably.Abstractions.Testing.Tests.Internal;
+namespace Testably.Abstractions.Testing.Tests.Helpers;
 
 public class FilePlatformIndependenceExtensionsTests
 {
