+++ Tests/Testably.Abstractions.Testing.Tests/Helpers/FileSystemExtensionsTests.cs
@@ -1,6 +1,6 @@
 ﻿using Testably.Abstractions.Testing.Helpers;
 
-namespace Testably.Abstractions.Testing.Tests.Internal;
+namespace Testably.Abstractions.Testing.Tests.Helpers;
 
 public class FileSystemExtensionsTests
 {
