+++ Tests/Testably.Abstractions.Testing.Tests/Helpers/PathHelperTests.cs
@@ -4,7 +4,7 @@ using Testably.Abstractions.FileSystem;
 using Testably.Abstractions.Testing.Helpers;
 using Testably.Abstractions.Testing.Tests.TestHelpers;
 
-namespace Testably.Abstractions.Testing.Tests.Internal;
+namespace Testably.Abstractions.Testing.Tests.Helpers;
 
 public class PathHelperTests
 {
